refresh junit libs

+++ src/main/java/CheckPrime.java
@@ -2,7 +2,7 @@ import PrimeFunctions.*;
 
 public class CheckPrime {
     public static void main (String[] args) {
-        int numberIn = Integer.parseInt(args[0]);
+        Double numberIn = Double.parseDouble(args[0]);
         IsPrime isPrime = new IsPrime();
         boolean primeNumber = isPrime.IsPrime(numberIn);
